Update patch set 1
Patch Set 1:

This is the exact use case, and yeah, it's for experts only.  If you want it, you have to know how to format that partition - note that formatting support is not wired in.  This doesn't change ext4 to the default, it just makes it mountable.

The main upside is the instant mount on boot, plus the usual reasons to choose ext4.

App breakage, in practice, has not been a problem.  If anything quite the opposite - some got noticeably faster, less lag, etc.

Patch-set: 1