Update patch set 2
Patch Set 2:

This is the same test that is already done in Volume.cpp when formatting the volume and I also enabled in VolumeManager for sharing because we need the correct "/dev/block/vold/#major:#minor" when opening the lun file but the minor number gets reported incorrectly when the EMMC shares the major device number without this.

Sharing a major device number is perfectly valid, and this test should really be the default, but given that it isn't, this is needed for all volumes that share the major device number, as is true on Galaxy S phones for internal and external sd cards.

Patch-set: 2