Update patch set 1
Patch Set 1:

It's possible that it could choose an incorrect partition, but no more likely than vold doing so on it's own. Without this patch the format just fails, with the patch it tries a different method of obtaining the path before giving up which solves the issue.

Unfortunately I lack any additional hardware to test on, but if your phone can format the SD without issue, the code I've submitted won't even run. It's only running if vold has chosen a non-existent device.

I would, of course, appreciate any suggestions for improvements or if this is a bad idea, let me know and I'll scrap it =)

Patch-set: 1