Update patch set 3
Patch Set 3:

About the vold, someone is looking for this issue:
http://code.google.com/p/cyanogenmod/issues/detail?id=2949#c117

The USB write speed to SD card is extremely slow on the HTC Desire and to some other devices..

Maybe worth to look in this old commit:
https://github.com/CyanogenMod/android_system_vold/commit/a28056b38275003895ff5d9576681aca01544822

Is that correct to set the dirty_ratio to zero, when the sd card is mounted as mass storage? 

After reverting that commit the USB write speed to SD card its back to normal finally.

Patch-set: 3