Update patch set 2
Patch Set 2:

Judging by what people have written about the issue it seems to be kernels later than 2.6.35 in Gingbread that get bad performance if dirty_ratio is set to 0.

I did some tests on my ZTE V9 that runs 2.6.35 and the performance seems to be pretty much the same regardless if I set it to 0 or 20 (8Mb/s while copying a large file)

On my HTC Desire with kernel 2.6.37 I get 8Mb/s with dirty_ratio set to 20 and 1.1Mb/s when it's set to 0.

Patch-set: 2