Update patch set 1
Patch Set 1:

%as is a GNU extension -- see man fscanf:

An optional 'a' character.  This is used with string conversions, and relieves  the caller  of  the need to allocate a corresponding buffer to hold the input: instead, scanf() allocates a buffer of sufficient size, and assigns the address of this buffer  to  the  corresponding pointer argument, which should be a pointer to a char * variable (this variable does not need to be  initialized  before  the  call).   The caller should subsequently free(3) this buffer when it is no longer required.  This is a GNU extension; C99 employs the 'a' character as a conversion specifier (and it can also be used as such in the GNU implementation).

Patch-set: 1