Update patch set 1
Patch Set 1:

I'm not entirely sure this is sufficient, due to the new storage_list logic in ICS.

Also, this is a pretty straightforward patch from CM7, so I'd much rather see the original patch imported with full authorship and description rather than a generic "ported from CM7" blurb.

I'll give it a go in a few hours.

Patch-set: 1