Update patch set 1
Patch Set 1:

I resubmitted this because it was needed for es209ra when I did a build a few days ago. I see that at around the time I did that build a new prebuilt kernel was uploaded that  now supports NPARTS, but I think this patch is still good. It may no longer be needed by any officially supported CM devices (do we have anything else on 2.6.29?), but will probably be useful for someone downstream.

This patch will have no impact for devices that do supports NPARTS.

Patch-set: 1