Update patch set 2
Patch Set 2:

the code in question I'm talking about is in packages/apps/Settings/src/com/android/settings/deviceinfo/StorageMeasurement.java

        final File[] files = dir.listFiles();
        if (files == null) return 0;

        // Get sizes of all top level nodes except the ones already computed
        long counter = 0;
        long miscSize = 0;

        for (File file : files) {
            final String path = file.getAbsolutePath();
            final String name = file.getName();
            if (sMeasureMediaTypes.contains(name)) {
                continue;
            }

            if (file.isFile()) {
                final long fileSize = file.length();
                mFileInfoForMisc.add(new FileInfo(path, fileSize, counter++));
                miscSize += fileSize;
            } else if (file.isDirectory()) {
                final long dirSize = getDirectorySize(imcs, file);
                mFileInfoForMisc.add(new FileInfo(path, dirSize, counter++));
                miscSize += dirSize;
            } else {
                // Non directory, non file: not listed
            }
        }

In order to get the initial directory listing, the application needs to access the directory (and the DefaultContainerService provides no such functionality). It seems to me that it is proper that the app can access it because it has android.permission.READ_EXTERNAL_STORAGE in its AndroidManifest.xml. Note that on devices where emulated sdcard is used, the UID is, in fact, set to root, allowing the functionality to work properly.

Patch-set: 2