Update patch set 4
Patch Set 4:

Debian and other distros are including exfat-fuse. We may be even more in the clear as the license was already obtained by the device manufacturer. I think that I would rather have the interoperability now, and if the patent is actually enforced then we just remove it. And of course if someone ships a commercial product with it then its their problem.

Patch-set: 4