Update patch set 6
Patch Set 6: Verified+1 Code-Review+2

If it becomes an issue, we'll just remove it and find another solution. Debian's stance on the issue is that they aren't tied to it, and it's not their responsibility to be policing patents- it's the responsibility of whoever decides to use it for commercial purposes. I agree 100% and would rather have the interoperability so that people don't need to reformat all of their stuff. If later on we do anything commercial like building a phone, we will deal with it.

Patch-set: 6
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2