Update patch set 1
Patch Set 1: Code-Review-1

Hi Pawit i agree this should not be merged. The problem is not that we don't have a method to get the volume id. getVolumeId in this commit is copied directly from FileUtils. The issue is that MediaProvider can no longer read the volume id for primary (maybe secordary) removable storage. Since MediaProvider is passed /storage/sdcard0 FileUtils.getFatVolumeId always fails since its the fuse mount point. we can change it to read /mnt/media_rw/sdcard0 but it doesnt matter since its mounted 700 MediaProvider lacks the necessary permissions to read it. Furthermore (for the sake of testing) if we modify init to create /mnt/media_rw as 777 and modify vold to mount /mnt/media_rw/sdcard0 as 777 MediaProvider /still/ cant read it. My guess is zygote is blocking the jni call for getFatVolumeId but i don't know... at that point i gave up and did this.

Patch-set: 1
Label: Code-Review=-1