Update patch set 1
Patch Set 1:

Also... to address concerns raised (which really should be in the other ps)...

The code may easily support other file system types. This was pointed out already.

The file system id may be reset to -1 on unmount.

These are minor tweaks, not fundamental issues with the  implementation.

Patch-set: 1