Update patch set 1
Patch Set 1:

fwiw i tried again having it read /mnt/media_rw/sdcard0 and it works fine, not sure why i was having issues before. only change needed is to chmod 770 /mnt/media_rw. So really just need a way to find the real mount point and call getFatVolumeId/getVolumeUUID on it. But it would probably be the same procedure as pawit did to get the uuid.

Patch-set: 1