Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

In my code, I've stripped all the dashes and limited the hex to 8 characters in order to fit an int.

The ioctl for FAT can also be removed for code simplification since blkid also handle FAT volumes. This also ensures that the blkid code path actually works.

BTW, what's wrong with my patch set? IMO, it's a waste of time to be reviewing and fixing this if the other one works fine.

Patch-set: 3
Label: Code-Review=-1