Update patch set 3
Patch Set 3:

Can't remember latest working for me. 6th-10th January was ok for sure. I remember this because I used external usb with 3rd party camera app to record movie directly to usb storage. I'm trying to do this again on today nightly and it's not possible any longer. All apps are unable to write to usb storage, I need to edit platform xml, add permissions to make it working.

I believe it should not be needed, but it seems to be only solution in ma case.

Patch-set: 3