Update patch set 1
Patch Set 1:

Hi Rashed,

When installing this version on an S3 with no encryption (blank), trying to encrypt the device will generate a "Cannot seek to write persistent data". After adding the write_offset to the error message, I realized that:
- write_offset was not a 32b offset but a 64b offset
- write_offset was bigger than INT32_MAX and when truncated, resulted in a negative value (which doesn't work for lseek here).
As a result, the seek fails and the save_persistent_data fails (but the calling function doesn't check for the returned value and continues).

If you look at the "read" call just few lines before (https://github.com/CyanogenMod/android_system_vold/blob/cm-12.0/cryptfs.c#L873), it is using lseek64 and not lseek.

It is true that if "_FILE_OFFSET_BITS" is defined (which was not the case in my environment), both are equivalent. However I think that as the code is using off64_t and not off_t, it would be cleaner to also directly use the 64 variant of lseek. At the very least, the code should not alternate between lseek and lseek64.

Sorry for the low quality of my previous message,
Vincent

Patch-set: 1