Update patch set 1
Patch Set 1:

Context for reviewers: this is basically a different way to fix the same decryption failure for which I backported Id1761b9. CAF vold assumes the parser from Linux 3.1, which accepts 0 as meaning "0 extra arguments", but some CyanogenMod ports only have Linux 3.0 (the one I'm interested in is the Galaxy Nexus family).

This seems to me like more complexity than it's worth. If this is going to be addressed from the vold side at all, the minimal-complexity version would be to revert one line of commit db2b1edd:

-  sprintf(crypt_params, "%s %s 0 %s 0 %s 0", crypt_ftr->crypto_type_name,
+  sprintf(crypt_params, "%s %s 0 %s 0 %s", crypt_ftr->crypto_type_name,

but that would need to be tested on a device with a 3.1+ kernel, preferably one that uses the Inline Crypto Engine (ICE) for which the change was made in the first place. I don't have such a device, so I can't test that myself.

Patch-set: 1
Reviewer: Simon McVittie <14122@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0