Update patch set 2
Patch Set 2:

A fresh install of M works fine and encryption works.  If I don't set a PIN or password before encryption, it will use the default password and setting a password or PIN after still doesn't prompt at boot.  Not directly related to this change, but I think the same root issue is causing that problem as my migration to M.

Steve, was this only to  correct OTA issues with a default password?

Patch-set: 2