Update patch set 1
Patch Set 1:

We do not build the hw part in cryptfs as our build flag is ignored as we're after include $(CLEAR_VARS) currently. We  messed up in an earlier patch.

Quick check: place bogus characters inside a CONFIG_HW_DISK_ENCRYPTION ifdef of your choice in cryptfs and mmm -B vold.

Whatever with the tabs, I don't care as long is it works.

Patch-set: 1