Update patch set 1
Patch Set 1:

Hi Lukasz,

No, this is just for the vold crash issue Paul mentioned here: http://review.cyanogenmod.org/#/c/126216/

On checking taking a long time - that's something I've noticed too (which I'm not sure has anything to do with cm).  What I'm seeing is that when doing a upgrade to a new build, f2fs.fsck takes a /lot/ longer to run (like 1-2mins instead of 5s).  I don't think it's a big problem but would be interesting to get to the bottom of.

Patch-set: 1