Update patch set 2
Patch Set 2:

@Ethan,

On bacon when encrypting, the device only soft reboots which leaves some leftover cruft causing the decryption to fail until the device is rebooted.

This gives the users an impression there is an issue with decrypting even though it's just that the encrypted device was still mounted somehow.

Patch-set: 2