Update patch set 1
Patch Set 1:

@Angus indeed nice bug report.  This looks like a real problem.

@Gabriele I agree it would be nicer for vold to restorecon directly.  However, vold probably doesn't run restorecon directly because of selinux permissions.

Comments in the vold commits indicate that this (setting a property and letting init run restorecon) is how /data is done.  so would this problem also affect /data if it had too many files?

I think a better solution would be to have init run restorecon_recursive asynchronously.  Either in a thread or in a fork.  init doesn't seem to use threads, so fork may by the way to go.

Patch-set: 1