Update patch set 1
Patch Set 1:

Why would this be needed only for hardware-based encryption? It doesn't seem obvious to me, I went through the code and the value of the variables you moved shouldn't depend on the type of encryption. They seem to be set by cryptfs_check_passwd(), which should be called by "vdc cryptfs checkpw <password>".

Patch-set: 1