Update patch set 1
Patch Set 1:

> (1 comment)

Because doing that check where you suggested requires not only adding an ifdef, but also another call to is_hw_disk_encryption, and in any case, structuring it to avoid the call to get_crypt_ftr_and_key is not important since both checks (master_key_saved and saved_mount_point) are highly likely to NOT lead to the abort flow (the code looks like they are almost assertion-level checks).

Also, I specifically moved the structure of the ifdefs the way I did. A lot of blocks in the android platform code are written as:
#ifdef SOME_FEATURE_ENABLED
if (is_some_feature_enabled()) { 
    ... implementation that uses feature ...
} else {
    ... fallback implementation ...
}
#else
... fallback implementation (copy pasted from before) ...
#endif

This code duplication is extremely bad practice, and is highly likely to bite us one day when someone changes one piece of code and forgets to change the other. The way I structured it is one option to avoid duplication; Another good option is to move the IFDEF inside is_hw_disk_encryption and make it always return FALSE if CONFIG_HW_DISK_ENCRYPTION is not defined (note that this does not degrade performance - due to compiler optimizations, if CONFIG_HW_DISK_ENCRYPTION is not defined, the whole function will be inlined as FALSE, and all conditionals statements invoking it will become if(FALSE) and be optimized out at compile time).

In any case, those are just my 2 cents on the issue (which is completely unrelated to the bug at hand, by the way); I do not have the time nor am I familiar enough with the code and the developer base to change the core functionality of hw_disk_encryption like that. I just want this bug fixed so I can use verifypw on my phone when I'm hacking around cryptfs.

Patch-set: 1