Update patch set 1
Patch Set 1:

Would you mind adding a bit more info to the commit message so that it's a bit more clear why you are doing this without looking at the code?

Also, the title of the commit could be a bit shorter, you could just mention what you are fixing and then explain how in the body of the message.

Patch-set: 1