Revert^2 "Add @SensitiveData tag to IVold"
This reverts commit 697c6a52173abcb6106180c36b0b0574db21cb2c.
It is reapplying the change in I0439f63fd4739bf5a6c957695cc9c3003ec89eb0.

Reason for revert: Undoing the revert (putting the change back in); after looking at the 
performance bug, it seems impossible it was caused by the addition of the `@SensitiveData` 
tag on IVold.
Performance bug: 331045735

Bug: 320392352
Test: launch_cvd -daemon

Change-Id: I522f63836155ea404260e89fd2f209738f37d5b3